refactor: enhance pagination handling with generic listAllPaginated function
- Introduced generic listAllPaginated function to centralize pagination logic
- Updated list functions (listAll, teamFoldersAll, sharedWithMeAll) to use
  the new pagination helper
- Changed offset/limit parameters from int to int64 for consistency with API
  responses
- Improved error messages and code organization
- Updated mock implementations and tests to match the new function signatures

+++ synology_drive_api/list_test.go
@@ -10,8 +10,8 @@ import (
 func TestList(t *testing.T) {
 	tests := []struct {
 		name          string
-		offset        int
-		limit         int
+		offset        int64
+		limit         int64
 		expectedError string
 	}{
 		{"valid request", 0, 100, ""},
