+++ synology_drive_exporter/exporter_test.go
@@ -46,13 +46,13 @@ func (m *MockFileSystem) CreateFile(filename string, data []byte, dirPerm os.Fil
 }
 
 type MockSynologySession struct {
-	ListFunc         func(rootDirID synd.FileID, offset, limit int) (*synd.ListResponse, error)
+	ListFunc         func(rootDirID synd.FileID, offset, limit int64) (*synd.ListResponse, error)
 	ExportFunc       func(fileID synd.FileID) (*synd.ExportResponse, error)
-	TeamFolderFunc   func(offset, limit int) (*synd.TeamFolderResponse, error)
-	SharedWithMeFunc func(offset, limit int) (*synd.SharedWithMeResponse, error)
+	TeamFolderFunc   func(offset, limit int64) (*synd.TeamFolderResponse, error)
+	SharedWithMeFunc func(offset, limit int64) (*synd.SharedWithMeResponse, error)
 }
 
-func (m *MockSynologySession) List(rootDirID synd.FileID, offset, limit int) (*synd.ListResponse, error) {
+func (m *MockSynologySession) List(rootDirID synd.FileID, offset, limit int64) (*synd.ListResponse, error) {
 	if m.ListFunc == nil {
 		return nil, errors.New("ListFunc not set")
 	}
@@ -66,14 +66,14 @@ func (m *MockSynologySession) Export(fileID synd.FileID) (*synd.ExportResponse, 
 	return &synd.ExportResponse{}, nil
 }
 
-func (m *MockSynologySession) TeamFolder(offset, limit int) (*synd.TeamFolderResponse, error) {
+func (m *MockSynologySession) TeamFolder(offset, limit int64) (*synd.TeamFolderResponse, error) {
 	if m.TeamFolderFunc != nil {
 		return m.TeamFolderFunc(offset, limit)
 	}
 	return &synd.TeamFolderResponse{}, nil
 }
 
-func (m *MockSynologySession) SharedWithMe(offset, limit int) (*synd.SharedWithMeResponse, error) {
+func (m *MockSynologySession) SharedWithMe(offset, limit int64) (*synd.SharedWithMeResponse, error) {
 	if m.SharedWithMeFunc != nil {
 		return m.SharedWithMeFunc(offset, limit)
 	}
@@ -387,7 +387,7 @@ func TestExporterExportMyDrive(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			// Create mocks
 			mockSession := &MockSynologySession{
-				ListFunc: func(rootDirID synd.FileID, offset, limit int) (*synd.ListResponse, error) {
+				ListFunc: func(rootDirID synd.FileID, offset, limit int64) (*synd.ListResponse, error) {
 					// Track that this directory was listed
 					if tt.trackedListCalls != nil {
 						tt.trackedListCalls[rootDirID] = true
